+++ devHelper/plenTextEmailTemp.txt
